Update minmax files

+++ minmax/src/custom_minmax_vector.cpp
@@ -5,25 +5,29 @@ int main() {
     std::cout << "Minimum and maximum elements of vector\n";
     std::cout << "Custom implementation\n\n";
 
+    // Create vector
     std::vector<int> nums;
 
+    // Ask user for vector size
     std::cout << "How many integers? ";
     int size;
     std::cin >> size;
     std::cout << '\n';
 
-    for (size_t i = 0; i < size; ++i) {
+    // Ask user for integers
+    for (int i = 0; i < size; ++i) {
         std::cout << "Enter an integer: ";
         int num;
         std::cin >> num;
         nums.push_back(num);
     }
 
+    // Start at first element
     int min = nums.at(0);
     int max = nums.at(0);
 
     std::cout << "\n[";
-    for (size_t i = 0; i < nums.size(); ++i) {
+    for (int i = 0; i < nums.size(); ++i) {
         int num = nums.at(i);
 
         // Update `min` and `max`
